Update README.md
INET3101 LAB 2 Memory and Structures:
This program is a simple inventory tracker in C that helps you store information about parts, like their number, name, size, and cost.
You can choose from four options in the menu: add a new part, see all parts, check how many parts are stored, or exit the program. 
Each part is saved as a record with details such as part number, name, size, and cost. 
All the parts are stored in a list, and the program keeps track of how many parts are added.